crawler-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async HTTP
aiohttp==3.9.1
aiofiles==23.2.1

# Redis for queue management
redis==5.0.1
hiredis==2.2.3

# Bloom filter for duplicate detection
pybloom-live==4.0.0

# Robots.txt parsing
reppy==0.4.14
urllib3==2.1.0

# HTML parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# URL parsing and normalization
yarl==1.9.4
tldextract==5.1.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Rate limiting
aiolimiter==1.1.0

# Database
asyncpg==0.29.0
motor==3.3.2

# Process management
multiprocess==0.70.15
aiomultiprocess==0.9.0

# Sitemap parsing
ultimate-sitemap-parser==0.5

# DNS resolution
aiodns==3.1.1

# Content type detection
python-magic==0.4.27
chardet==5.2.0
langdetect==1.0.9
ftfy==6.1.3

# JavaScript rendering
playwright==1.40.0
playwright-stealth==1.0.6

# Structured data extraction
extruct==0.16.0
jsonschema==4.20.0
rdflib==7.0.0

# Content deduplication
simhash==2.1.2
datasketch==1.6.4
xxhash==3.4.1

# Image processing
Pillow==10.1.0

# Additional utilities
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2